Hyperspace jammer reverted to use Escape key from hyperspace
A J appears in the upper left of the space window if the jammer is active.
It goes away once you navigate again, but is still active. Next time you hit Escape,
the J should disappear. This is a short-term position and key solution. Anyone
with code suggestions on a better way to add a key or a better indicator and location,
please let me know.